Following MVC pattern
