+++ QuiRky/app/src/main/AndroidManifest.xml
@@ -39,7 +39,8 @@
         <activity
             android:name=".MapActivity"
             android:exported="false"
-            android:label="@string/title_activity_map" />
+            android:label="@string/title_activity_map"
+            android:parentActivityName=".StartingPageActivity"/>
         <activity
             android:name=".StartingPageActivity"
             android:exported="false"
